requirements.txt
scrapy-playwright==0.0.26
fastapi==0.100.1
uvicorn==0.23.2
python-dotenv==1.0.0
# psycopg2==2.9.6
psycopg2-binary